flask_poe_back_end/requirements.txt
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
Flask==3.0.0
google-ai-generativelanguage==0.3.3
google-api-core==2.12.0
google-auth==2.23.3
google-generativeai==0.2.1
googleapis-common-protos==1.61.0
grpcio==1.59.0
grpcio-status==1.59.0
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
proto-plus==1.22.3
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
requests==2.31.0
rsa==4.9
tqdm==4.66.1
urllib3==2.0.6
Werkzeug==3.0.0